moved SymWorld.h to source/default_mode/ and removed WorldSetup.cc from source since each mode now has its own WorldSetup file
