+++ source/default_mode/SymWorld.h
@@ -1,11 +1,11 @@
 #ifndef SYM_WORLD_H
 #define SYM_WORLD_H
 
-#include "../../Empirical/include/emp/Evolve/World.hpp"
-#include "../../Empirical/include/emp/data/DataFile.hpp"
-#include "../../Empirical/include/emp/math/random_utils.hpp"
-#include "../../Empirical/include/emp/math/Random.hpp"
-#include "Organism.h"
+#include "../../../Empirical/include/emp/Evolve/World.hpp"
+#include "../../../Empirical/include/emp/data/DataFile.hpp"
+#include "../../../Empirical/include/emp/math/random_utils.hpp"
+#include "../../../Empirical/include/emp/math/Random.hpp"
+#include "../Organism.h"
 #include <set>
 #include <math.h>
